+++ src/test/java/io/github/astrapi69/lang/DefaultValueTest.java
@@ -20,13 +20,10 @@
  */
 package io.github.astrapi69.lang;
 
-import static org.testng.AssertJUnit.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
-import java.lang.reflect.InvocationTargetException;
-
-import org.meanbean.test.BeanTestException;
+import org.junit.jupiter.api.Test;
 import org.meanbean.test.BeanTester;
-import org.testng.annotations.Test;
 
 /**
  * The unit test class for the class {@link DefaultValue}.
@@ -86,8 +83,7 @@ public class DefaultValueTest
 	/**
 	 * Test method for {@link DefaultValue} with {@link BeanTester}
 	 */
-	@Test(expectedExceptions = { BeanTestException.class, InvocationTargetException.class,
-			UnsupportedOperationException.class })
+	@Test
 	public void testWithBeanTester()
 	{
 		final BeanTester beanTester = new BeanTester();